update to use System.Text.Json
While updating this tutorial, let's use the latest features in .NET Core 3.1
