+++ csharp/getting-started/console-webapiclient/Repository.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Globalization;
+using System.Text.Json.Serialization;
+
+namespace WebAPIClient
+{
+    public class Repository
+    {
+        [JsonPropertyName("name")]
+        public string Name { get; set; }
+        [JsonPropertyName("description")]
+        public string Description { get; set; }
+
+        [JsonPropertyName("html_url")]
+        public Uri GitHubHomeUrl { get; set; }
+
+        [JsonPropertyName("homepage")]
+        public Uri Homepage { get; set; }
+
+        [JsonPropertyName("watchers")]
+        public int Watchers { get; set; }
+
+        [JsonPropertyName("pushed_at")]
+        public string JsonDate { get; set; }
+        
+        public DateTime LastPush =>
+            DateTime.ParseExact(JsonDate, "yyyy-MM-ddTHH:mm:ssZ", CultureInfo.InvariantCulture);
+    }
+}
