+++ csharp/getting-started/console-webapiclient/webapiclient.csproj
@@ -1,7 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
    <ItemGroup>
-    <PackageReference Include="System.Runtime.Serialization.Json" Version="4.3.0" />
-    <PackageReference Include="System.Runtime.Serialization.Primitives" Version="4.3.0" />
+     <PackageReference Include="System.Text.Json" Version="4.7.0" />
    </ItemGroup> 
 
   <PropertyGroup>
